requirements.txt
Django==2.2
django-filter==2.3.0
djangorestframework==3.10.3
Markdown==3.2.2
django-cors-headers==3.1.1
factory_boy==2.12.0